Add font folder and apply "Poppins" font to PropertyCard component.
Create "assets" folder in the project root directory.
Inside the "assets" folder, create a "fonts" subfolder.
Download the "Poppins-Regular.ttf" font file and place it in the "fonts" folder.
Create a "fonts.js" file in the root directory.
Import the "expo-font" package in "fonts.js" to load fonts.
Define a "loadFonts" function in "fonts.js" to load the "Poppins" font.
Modify the "Card.jsx" component file to import the "useFonts" hook from "expo-font".
Use the "useFonts" hook to load the "Poppins" font in the "PropertyCard" component.
Apply the "Poppins" font to the appropriate text elements in the "PropertyCard" component's styles.