Update the example screenshot of functional app to README
In this commit, I have added a new screenshot of the functional app as an example to the README file. This screenshot provides a visual representation of the app in action, allowing users to get a glimpse of its functionality. It enhances the README's content and provides a better understanding of what the app offers.